common: Prevent CONTEXT corruption by StackWalk64.
StackWalk64 modifies the received context, which creates problems when
attempting to resume execution.  This is a fact I keep forgetting, so
this change makes it hard to repeat the error yet again.

This also explains why the minidump exception records were getting
corrupt -- because the dump was being done after StackWalk64 call.