cmake: bug fix - refine best guess of file name to dlltool
Replace either 'gcc' and - new - 'cc' in the file name of the C compiler
determined by cmake to come up with a best guessed file name of the
dlltool.

Fixes the bug when the file name of the C compiler does not contain the
string 'gcc', but 'cc' only  (observed case on MSYS2/MinGW64 where cmake
determined the C compiler file name as 'cc.exe').