Correctly handle Unicode file names in ExcHndl
Add `ExcHndlSetLogFileNameW` that accepts a Unicode wide character string as an alternative to `ExcHndlSetLogFileNameA`, which uses the current Windows code page.
Internally only the wide character string is stored. The file name passed to `ExcHndlSetLogFileNameA` is converted to wide characters using the current code page.

Use `GetModuleFileNameW` instead of `GetModuleFileNameA` to handle the case that the module file name contains Unicode characters.

Don't use `%s` in format strings to print the argument when it's a nullptr, as this is undefined behavior.