Extend ExcHndl tests for `ExcHndlSetLogFileNameW`
As a lot of testing code also had to be adjusted for usage of wide character strings, it was easiest to split the existing test code by adding another compile time flag `TEST_UNICODE`.
The flag is not named `UNICODE`, as this name is already used by the Windows API.

A macro `WFILE` is added that expands to a wide character string literal containing `__FILE__`.

The output file name for the unicode tests contains additional unicode characters to properly test the behavior.